Add "zone" as a supported format for mlabconfig.
This change adds a new output format for mlabconfig to generate a zone
file for measurement-lab.org. This form of generation should obsolete
gen-zone.py.

The mlabzone.header.in is taken from the current gen-zone.py script. The
included version of mlabzone.header.in increases the zone expiration
time from 1 day to 7 days. This should allow for longer periods of
outages of the shadow master server.